last build

+++ src/main/java/com/makarov/lshop/base/dao/api/repository/ProfileRepository.java
@@ -3,9 +3,7 @@ package com.makarov.lshop.base.dao.api.repository;
 import com.makarov.lshop.base.model.ProfileEntity;
 import org.springframework.data.repository.CrudRepository;
 
-import java.math.BigInteger;
-
-public interface ProfileRepository extends CrudRepository<ProfileEntity, BigInteger> {
+public interface ProfileRepository extends CrudRepository<ProfileEntity, Long> {
 
     ProfileEntity findByLogin(String login);
 }
